jira-oauth-token/requirements.txt
tlslite-ng==0.7.5
oauth2==1.9.0.post1
